requirements.txt
openai-whisper==20231117
flask==3.0.0
flask-cors==4.0.0
razorpay==1.4.1
torch==2.1.0